==> 04_Figure_1_CD.Rmd <==
---
title: "A novel role for alkyladenine DNA glycosylase in regulating alkylation-induced ER stress"
subtitle: "Figure 1 C and D"
author: "L Milano, CF Charlier, R Andreguetti, E Healing, MP Thomé, R Elliott, JY Masson, LD Samson, G Lenz, JAP Henriques, A Nohturfft and LB Meira"
output:
  html_notebook:
    toc: TRUE
    toc_float: TRUE
    toc_depth: 3
    code_folding: "hide"
    number_sections: FALSE
    theme: "readable"
    highlight: "tango"
    fig_caption: TRUE
    css: "./source/styles.css"
---

```{js ShowHide Session Info}
function myFunction(id) {
  var x = document.getElementById(id);
  if (x.style.display === 'none') {
    x.style.display = 'block';
  } else {
    x.style.display = 'none';
  }
}

```


<div class='comments'>
Note: This analysis requires input data generated by an earlier script.
</div>

## Setup  

* Clear memory and console  
* Load packages  
* Get current script name  

```{r CLEAR MEMORY AND PACKAGES AND CONSOLE, results="hide"}
# Clear memory
rm(list=setdiff(ls(all=TRUE), c(".Random.seed")))
# Clear console
cat("\014")
```


```{r load packages, include=FALSE}
library(magrittr)
library(dplyr)
library(ggplot2)
```


```{r class.output="txt_output"}
this.script <- rstudioapi::getActiveDocumentContext()$path %>% basename
cat("Script:", this.script)
```

## Load data  
### Load toptables generated in previous script:  
```{r results="hide"}
xlsx.file <- "Supplemental_Data_1.xlsx"
stopifnot(file.exists(xlsx.file))
sheet.names <- openxlsx::getSheetNames(xlsx.file)
tt <- lapply(sheet.names, function(s.name) {
   res <- openxlsx::read.xlsx(xlsxFile = xlsx.file, sheet=s.name) %>% 
     dplyr::select(SYMBOL, logFC, adj.P.Val) %>% 
     dplyr::filter(complete.cases(.)) %>% 
     dplyr::group_by(SYMBOL) %>% 
     dplyr::slice(base::which.min(adj.P.Val)) %>% 
     dplyr::ungroup(.)
   res
}) %>% 
  set_names(sheet.names)
```

```{r class.output="txt_output"}
cat("File read:", xlsx.file)
rm(xlsx.file)
```


### Load mouse/human gene symbol map  
```{r results="hide"}
map.file <- "./data/mouse_human_symbols_2018_03_05b.txt.gz"
stopifnot(file.exists(map.file))
gs.ms.hu <- read.table(file=map.file, sep="\t", header=TRUE,
                       stringsAsFactors = FALSE)
head(gs.ms.hu)
```


### Load gene ontology (GO) data  
The original input file was obtained from the [enrichr](https://maayanlab.cloud/Enrichr/) website ([Chen et al., 2013](https://pubmed.ncbi.nlm.nih.gov/23586463/); [Kuleshov et al., 2016](https://pubmed.ncbi.nlm.nih.gov/27141961/)).

```{r class.output="txt_output"}
list.files("data")
enrichr.GO_Biological_Process.file <- "./data/GO_Biological_Process_2015.txt.gz"
stopifnot(file.exists(enrichr.GO_Biological_Process.file))
search.pattern <- "endoplasmic reticulum unfolded protein response \\(GO:0030968\\)"
go.0030968.hu <- readLines(con=enrichr.GO_Biological_Process.file) %>% 
  grep(pattern=search.pattern, x=., invert = FALSE, value=TRUE) %>% 
  strsplit(., split="\t") %>% .[[1]] %>% .[-(1:2)]
cat("File read:", enrichr.GO_Biological_Process.file)
rm(enrichr.GO_Biological_Process.file, search.pattern)
```

```{r map to mouse symbols, class.output="txt_output"}
go.0030968.ms <- subset(gs.ms.hu, human %in% go.0030968.hu) %>%
  .$mouse %>% unique %>% sort
cat("Number of GO:0030968 genes:", length(go.0030968.ms))
rm(go.0030968.hu)
```


### Load Xbp1 ChIP-seq data  
The Xbp1 ChIP-seq targets were published by Argemí et al.(2017) ([Pubmed](https://www.ncbi.nlm.nih.gov/pubmed/28082079)) ([GEO](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE86048)).  
Data were downloaded from [ChIP-Atlas](http://chip-atlas.org/) and saved as a tsv file.  

Sample information was obtained from ChIP-Atlas:  
```{r}
argemi.samples <- data.frame(
  SRX = c("SRX2053270", "SRX2053271",
          "SRX2053272", "SRX2053273",
          "SRX2053274", "SRX2053275",
          "SRX2053276", "SRX2053277"),
  GSM = c("GSM2292554", "GSM2292555",
          "GSM2292556", "GSM2292557",
          "GSM2292558", "GSM2292559",
          "GSM2292560", "GSM2292561"),
  source_name = c("normal liver",
                  "liver: 7d after XBP1 silencing",
                  "liver: 6h after partial hepatectomy",
                  "liver: 7d after XBP1 silencing and 6h after partial hepatectomy",
                  "liver: 48h after partial hepatectomy",
                  "liver: 7d after XBP1 silencing and 48h after partial hepatectomy",
                  "liver: 6h after sham operation",
                  "liver: 48h after sham operation")
)
argemi.samples$GEO <- paste0("<a href='https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=", argemi.samples$GSM, "'>",
                             argemi.samples$GSM, "</a>")
argemi.samples %>% dplyr::select(SRX, GEO, source_name) %>% 
  DT::datatable(., rownames = F, options = list(dom = 't'), escape=F)
```

Download ChIP-seq data at maximum 1kb from transcription start site:  
```{r class.output="txt_output"}
xbp1.chip.seq.file <- "./data/chip_atlas_XBP1_mouse_1kb.tsv.gz"
stopifnot(file.exists(xbp1.chip.seq.file))
chip.atlas.1kb <- read.table(file=xbp1.chip.seq.file, sep="\t", header=T, stringsAsFactors = F)
cat("File read:", xbp1.chip.seq.file)
names(chip.atlas.1kb) <- sub("\\.Liver$", "", names(chip.atlas.1kb))
rm(xbp1.chip.seq.file)
```


```{r results="hide", include=FALSE}
# What does the ChIP-Atlas table look like (table consists of q-values,
# which represent binding significance as -log10 FDR-corrected p values):  
head(chip.atlas.1kb[,1:4])
```

Select control samples and those where ER stress was induced through partial hepatectomy:  
```{r}
chip.samples <- c("SRX2053270", "SRX2053272", "SRX2053274")
argemi.samples.select.1 <- argemi.samples %>%
  dplyr::select(SRX, GEO, source_name) %>% 
  dplyr::filter(SRX %in% chip.samples)
DT::datatable(argemi.samples.select.1, rownames = F, options = list(dom = 't'), escape=F)
```

ChIP-seq filtering parameters:  
```{r results="asis", class.output="txt_output"}
q.min <- 2000
cat("Maximum q-value: <b>", q.min, "</b><br>\n", sep="")
```


Number of target genes in each of the 3 samples:
```{r}
chip.atlas.1kb.select <- chip.atlas.1kb[,c("Target_genes", chip.samples)]
tmp <-sapply(chip.samples, function(srx) {
  chip.atlas.1kb.select[,c("Target_genes", srx)] %>% 
  magrittr::set_colnames(c("genes", "sample")) %>% 
  dplyr::filter(sample >= q.min) %>%
    .$genes %>% unique %>% length
})
chip.atlas.1kb.select.df <- data.frame(SRX=names(tmp), target.count=tmp)
argemi.samples.select.2 <- merge(x=argemi.samples.select.1, y=chip.atlas.1kb.select.df, by="SRX")
DT::datatable(argemi.samples.select.2, rownames = F, options = list(dom = 't'), escape=F)
rm(tmp, chip.atlas.1kb.select.df)
```

Select Xbp1 targets identified in any of the three samples:  
```{r results="asis"}
xbp1.chipseq.targets <- lapply(chip.samples, function(srx, cutoff.q, chip.atlas.1kb) {
  chip.atlas.1kb[, c("Target_genes", srx)] %>% 
    magrittr::set_colnames(c("genes", "sample")) %>% 
    dplyr::filter(sample >= q.min) %>% 
    .$genes %>% unique %>% sort
}, cutoff.q=cutoff.q, chip.atlas.1kb=chip.atlas.1kb) %>% 
  unlist %>% unique %>% sort
cat("Number of Xbp1 targets selected:<b>", length(xbp1.chipseq.targets), "</b><br>\n")
```

 
### Load Xbp1 correlation network data from Genecan  
These data were described in [Brignull et al., 2013](https://doi.org/10.1186/1471-2164-14-853).  
Genecan website: [http://genecan.sgul.ac.uk/](http://genecan.sgul.ac.uk/).  
Selecting 100 most strongly correlated genes.  
```{r class.output="txt_output"}
genecan.file.mouse.Xbp1 <- "./data/rMean_GPL1261_cvmin5_19Apr2012_1334842595_Xbp1.txt.gz"
stopifnot(file.exists(genecan.file.mouse.Xbp1))
xbp1.network <- read.table(file=genecan.file.mouse.Xbp1, sep="\t", header=T, stringsAsFactors = F) %>% 
    dplyr::arrange(desc(r_Mean)) %>% 
    head(., 100) %>% 
    .$Gene_Symbol
cat("File read:", genecan.file.mouse.Xbp1)
rm(genecan.file.mouse.Xbp1)
```

### Combine genesets to be highlighted  
```{r}
# Combine genesets in a list object: 
genesets <- list()
genesets$Xbp1.ChIPseq <- list(name="Xbp1 ChpSq", genes=xbp1.chipseq.targets)
genesets$GO_0030968 <- list(name ="GO:0030968", genes=go.0030968.ms)
genesets$Xbp1.Network <- list(name="Xbp1 Netwrk", genes=xbp1.network)
str(genesets)
```

## Volcano plots  
```{r}
source("./source/volcano_plot.source.R")
```


### Plot parameters    
```{r results="asis"}
# Plotting parameters volcano 1
plot.par <- list(
  font.size = 8,
  bar.height = 0.4,
  gap.height = 0.1,
  x.max = max(c(abs(tt$wt.mms.vs.wt.ctrl$logFC), abs(tt$ko.mms.vs.ko.ctrl$logFC))), # 4.957504
  y.max = max(c(-log10(tt$wt.mms.vs.wt.ctrl$adj.P.Val), -log10(tt$wt.mms.vs.wt.ctrl$adj.P.Val))),
  highlight.colors = c("#cc1d6f", RColorBrewer::brewer.pal(9, "Set1")),
  dot.sizes = c(0.8, 0.8, 1.2, 1.2),
  alphas = rep(1, 4),
  cutoff.fc = log2(1.75),
  cutoff.adj.p = 0.05,
  colors = c("#caf2a7", "#6fcc1d", "#d4b2c2",  "#cc1d6f") # 3= "gray70"
)
details::details(plot.par, summary="Show plot parameters", lang=NULL)
```


### Plot panel C  
```{r}
vp1 <- my.volcano.plot(
  dafra=tt$wt.mms.vs.wt.ctrl,
  highlight=genesets,
  title="C. Aag Wild-type",
  subtitle="Effect of MMS",
  x.max=plot.par$x.max,
  y.max=plot.par$y.max,
  bar.height=plot.par$bar.height,
  gap.height=plot.par$gap.height,
  y.intercept=-0.1,
  y.position="left",
  font.size=plot.par$font.size,
  dot.sizes = plot.par$dot.sizes,
  highlight.colors=plot.par$highlight.colors,
  alphas = plot.par$alphas,
  cutoff.fc = plot.par$cutoff.fc,
  cutoff.adj.p = plot.par$cutoff.adj.p,
  colors = plot.par$colors
  )
vp1
```

### Plot panel D  
```{r}
vp2 <- my.volcano.plot(
  dafra=tt$ko.mms.vs.ko.ctrl,
  highlight=genesets,
  title="D. Aag Knockout",
  subtitle="Effect of MMS",
  x.max=plot.par$x.max,
  y.max=plot.par$y.max,
  bar.height=plot.par$bar.height,
  gap.height=plot.par$gap.height,
  y.intercept=-0.1,
  y.position="right",
  font.size=plot.par$font.size,
  dot.sizes = plot.par$dot.sizes,
  highlight.colors=plot.par$highlight.colors,
  alphas = plot.par$alphas,
  cutoff.fc = plot.par$cutoff.fc,
  cutoff.adj.p = plot.par$cutoff.adj.p,
  colors = plot.par$colors
  )
vp2
```



### Panels combined  
```{r}
gridExtra::grid.arrange(vp1, vp2, nrow=1)
```

```{r results="asis"}
pdf.file.volcano <- "Figure_1_CD.pdf"
pdf(file=pdf.file.volcano, paper="a4", width=11.3/2.54, height=8/2.54, useDingbats=F)
gridExtra::grid.arrange(vp1, vp2, nrow=1)
dev.off()
```

```{r class.output="txt_output"}
cat("Saved:", pdf.file.volcano)
```


<hr>

## Session info  

<button class="button" onclick="myFunction('DIV_5')">Show/hide session info</button>
<div id="DIV_5" class="div_default_hide">

```{r SESSION INFO DATE, results="asis"}
cat("Date:", format(Sys.time(), "%a %d-%b-%Y %H:%M:%S"), "<br>\n")
```

```{r print_session_info, R.options=list(width=70)}
devtools::session_info()
```
</div>
  
```{js}
var divsToHide = document.getElementsByClassName("div_default_hide");
for(var i = 0; i < divsToHide.length; i++)
{
  divsToHide[i].style.display = 'none';
}
```
